requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
datasets
matplotlib
pillow